#42 Refactor tests of lookup implementation with black-box approach
